+++ FrontEnd/src/component/Main.css
@@ -1,51 +1,124 @@
-.Status {
-  position: absolute;
-  left: -0.18%;
-  right: 71.18%;
-  height: 120px;
-  width: 322.44px;
-  top: 0%;
-  bottom: 85.85%;
-  font-size: 34px;
-  line-height: 160px;
-  background: rgba(33, 33, 33, 0.72);
-
-  text-align: center;
-}
-.Home {
-  background-color: #2f3536;
-  min-height: 100vh;
+.Background {
+  height: 100vh;
   display: flex;
   flex-direction: column;
-
   font-size: calc(10px + 2vmin);
   color: white;
+  background: #2f3536;
+  overflow-x: hidden;
+  overflow-y: hidden;
+  position: relative;
+}
+
+.Status {
+  position: absolute;
+  height: 120px;
+  width: 320px;
+  top: 0%;
+  bottom: 85.85%;
+  background: rgba(33, 33, 33, 0.72);
 }
 
 .Nav {
   position: absolute;
-  left: -0.18%;
-  bottom: 0%;
   height: 70px;
-  width: 322.44px;
-  margin-bottom: 0px;
-
-  background: rgba(33, 33, 33, 0.72);
+  width: 320px;
+  bottom: 0%;
+  background: #252727;
 }
 
 .Divider {
   position: absolute;
-  margin-left: 322.44px;
-  top: 0%;
-  bottom: 0%;
-  height: auto;
   width: 0.5px;
-
+  height: 100vh;
+  top: 0%;
+  left: 320px;
+  bottom: 0%;
   background: #8e8e93;
 }
 
-#components-layout-demo-fixed-sider .logo {
-  height: 32px;
-  background: rgba(255, 255, 255, 0.2);
-  margin: 16px;
+.Side {
+  position: absolute;
+  top: 120px;
+  left: 0px;
+}
+
+.Side-Account {
+  position: relative;
+  top: 220px;
+}
+
+.SideSearch {
+  position: absolute;
+  top: 0px;
+  left: 150px;
+}
+
+.searchBox {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, 50%);
+  background: #2f3640;
+  height: 40px;
+  border-radius: 40px;
+  padding: 10px;
+}
+
+.searchBox:hover > .searchInput {
+  width: 240px;
+  padding: 0 6px;
+}
+
+.searchBox:hover > .searchButton {
+  background: white;
+  color: #2f3640;
+}
+
+.searchButton {
+  color: white;
+  float: right;
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+  background: #2f3640;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  transition: 0.4s;
+}
+
+.searchInput {
+  border: none;
+  background: none;
+  outline: none;
+  float: left;
+  padding: 0;
+  color: white;
+  font-size: 16px;
+  transition: 0.4s;
+  line-height: 40px;
+  width: 0px;
+}
+
+@media screen and (max-width: 620px) {
+  .searchBox:hover > .searchInput {
+    width: 150px;
+    padding: 0 6px;
+  }
+}
+
+.FixedHeightContainer {
+  position: absolute;
+  top: 200px;
+  float: right;
+  width: 320px;
+  padding: 3px;
+}
+
+.Categories {
+  font-size: 50px;
+  height: 450px;
+  overflow: auto;
+  color: #000;
 }
