+++ FrontEnd/src/component/Main.js
@@ -2,71 +2,55 @@ import React, { Component } from 'react';
 import 'antd/dist/antd.css';
 import './Main.css';
 
-import { Layout, Menu, Icon } from 'antd';
-import Player from './Player';
-
-const { Header, Content, Footer, Sider } = Layout;
-
-export default class Main extends Component {
-  render() {
-    return (
-      <Layout>
-        <Sider
-          style={{
-            overflow: 'auto',
-            height: '100vh',
-            position: 'fixed',
-            left: 0
-          }}
-        >
-          <div className="logo" />
-          <Menu theme="dark" mode="inline" defaultSelectedKeys={['1']}>
-            <Menu.Item key="1">
-              <Icon type="user" />
-              <span className="nav-text">nav 1</span>
-            </Menu.Item>
-            <Menu.Item key="2">
-              <Icon type="video-camera" />
-              <span className="nav-text">nav 2</span>
-            </Menu.Item>
-            <Menu.Item key="3">
-              <Icon type="upload" />
-              <span className="nav-text">nav 3</span>
-            </Menu.Item>
-            <Menu.Item key="4">
-              <Icon type="bar-chart" />
-              <span className="nav-text">nav 4</span>
-            </Menu.Item>
-            <Menu.Item key="5">
-              <Icon type="cloud-o" />
-              <span className="nav-text">nav 5</span>
-            </Menu.Item>
-            <Menu.Item key="6">
-              <Icon type="appstore-o" />
-              <span className="nav-text">nav 6</span>
-            </Menu.Item>
-            <Menu.Item key="7">
-              <Icon type="team" />
-              <span className="nav-text">nav 7</span>
-            </Menu.Item>
-            <Menu.Item key="8">
-              <Icon type="shop" />
-              <span className="nav-text">nav 8</span>
-            </Menu.Item>
-          </Menu>
-        </Sider>
-        <Layout style={{ marginLeft: 200 }}>
-          <Header style={{ background: '#fff', padding: 0 }} />
-          <Content style={{ margin: '24px 16px 0', overflow: 'initial' }}>
-            <div
-              style={{ padding: 24, background: '#fff', textAlign: 'center' }}
-            ></div>
-          </Content>
-          <Footer style={{ textAlign: 'center' }}>
-            <Player />
-          </Footer>
-        </Layout>
-      </Layout>
-    );
-  }
+function Main() {
+  return (
+    <div className="Background">
+      <link
+        href="https://fonts.googleapis.com/icon?family=Material+Icons"
+        rel="stylesheet"
+      />
+      <div className="Side">
+        <div className="SideSearch">
+          <div className="searchBox">
+            <input
+              className="searchInput"
+              type="text"
+              name=""
+              placeholder="Search"
+            />
+            <button className="searchButton" href="#">
+              <i className="material-icons">search</i>
+            </button>
+          </div>
+        </div>
+        <div class="FixedHeightContainer">
+          <div class="Categories">
+            blah... ..blah blah blah... ..blah blah blah... ..blah blah blah...
+            ..blah blah blah... ..blah blah blah... ..blah blah blah... ..blah
+            blah blah... ..blah blah blah... ..blah blah blah... ..blah blah
+            blah... ..blah blah blah... ..blah blah blah... ..blah blah blah...
+            ..blah blah blah... ..blah blah blah... ..blah blah blah... ..blah
+            blah blah... ..blah blah blah... ..blah blah blah... ..blah blah
+            blah... ..blah blah blah... ..blah blah blah... ..blah blah blah...
+            ..blah blah blah... ..blah blah blah... ..blah blah blah... ..blah
+            blah blah... ..blah blah blah... ..blah blah blah... ..blah blah
+            blah... ..blah blah blah... ..blah blah blah... ..blah blah blah...
+            ..blah blah blah... ..blah blah blah... ..blah blah blah... ..blah
+            blah blah... ..blah blah blah... ..blah blah blah... ..blah blah
+            blah... ..blah blah blah... ..blah blah blah... ..blah blah blah...
+            ..blah blah blah... ..blah blah blah... ..blah blah blah... ..blah
+            blah blah... ..blah blah blah... ..blah blah blah... ..blah blah
+            blah... ..blah blah blah... ..blah blah blah... ..blah blah blah...
+            ..blah blah blah... ..blah blah blah...
+          </div>
+        </div>
+      </div>
+      <div className="Side-Account">Account Profile</div>
+      <div className="Status" />
+      <div className="Nav" />
+      <div className="Divider" />
+    </div>
+  );
 }
+
+export default Main;
